requirements.txt
Django
facepy
python-social-auth
django-discover-runner
coverage
mock
exam
